+++ benches/package.json
@@ -6,7 +6,7 @@
 	"description": "Benchmarks for Preact",
 	"scripts": {
 		"prepare": "cd proxy-packages/preact-v8-proxy && npm ci",
-		"start": "node ./scripts config text_update.html && tach --config dist/text_update.config.json --manual",
+		"start": "node ./scripts config many_updates.html && tach --config dist/many_updates.config.json --manual",
 		"bench": "node ./scripts bench",
 		"deopts": "node ./scripts deopts",
 		"help": "node ./scripts --help"
@@ -22,7 +22,7 @@
 		"globby": "^11.0.0",
 		"sade": "^1.7.3",
 		"strip-ansi": "^6.0.0",
-		"tachometer": "^0.5.1",
-		"v8-deopt-viewer": "^0.1.4"
+		"tachometer": "^0.5.5",
+		"v8-deopt-viewer": "^0.2.1"
 	}
 }
