+++ benches/src/02_replace1k.html
@@ -5,6 +5,14 @@
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<title>replace all rows</title>
 		<meta name="description" content="updating all 1,000 rows" />
+		<style>
+			.preloadicon {
+				display: none;
+			}
+			.glyphicon-remove:before {
+				content: '⨯';
+			}
+		</style>
 	</head>
 	<body>
 		<div id="main"></div>
@@ -14,19 +22,24 @@
 				measureMemory,
 				testElementText,
 				afterFrame,
-				afterFrameAsync
+				afterFrameAsync,
+				markRunStart,
+				markRunEnd
 			} from './util.js';
 			import * as framework from 'framework';
 			import { render } from '../src/keyed-children/index.js';
 
-			const { run } = render(framework);
+			const { run } = render(framework, document.getElementById('main'));
 
 			async function main() {
 				const elementSelector = 'tr:first-child > td:first-child';
 
+				// MUST BE KEPT IN SYNC WITH WARMUP COUNT IN benches/scripts/config.js
 				const WARMUP_COUNT = 5;
 				for (let i = 0; i < WARMUP_COUNT; i++) {
+					markRunStart(`warmup-${i}`);
 					run();
+					await markRunEnd(`warmup-${i}`);
 
 					await afterFrameAsync();
 					testElementText(elementSelector, (i * 1000 + 1).toFixed());
@@ -35,18 +48,20 @@
 				await afterFrameAsync();
 
 				afterFrame(function () {
-					testElementText(elementSelector, '5001');
+					testElementText(elementSelector, WARMUP_COUNT + '001');
 					performance.mark('stop');
 					performance.measure(measureName, 'start', 'stop');
 
 					measureMemory();
 				});
 
+				markRunStart('final');
 				performance.mark('start');
 				run();
+				await markRunEnd('final');
 			}
 
-			main();
+			afterFrame(main);
 		</script>
 	</body>
 </html>
